requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket support
websockets==12.0

# Twilio integration
twilio==8.10.0

# Database
supabase==2.0.3

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# SSL certificates
certifi==2023.11.17
